Try to get the docker hub doc correct.
Updating the doc on docker hub gets overwritten. It has been suggested that
the README.md overwrites. Of course the top level README.md has nothing in it
but one of the readme's next to the MD is reflected, the last one built. I'm
going to guess that I can put a README.md next to the base docker file and
have it apply to all the other images.