Use 'groupAccess' instead of chmod for changing SCC permissions
Using chmod in populate_scc.sh to change the permission causes all
the files in the directory in the lower layers of the docker images
to be duplicated in the current layer. This prevents sharing of the
common SCC layer between different application docker images when
they use this script to create application specific scc layer.
This change removes the use of chmod and uses 'groupAccess' sub-option
of -Xshareclasses to provide group read-write access to the cache.
However, this option requires umask to be set up properly (to 002) to be
able to set group read-write access on the SCC file.
More information about groupAccess sub-option can be seen at:
https://www.eclipse.org/openj9/docs/xshareclasses/#groupaccess

Signed-off-by: Ashutosh Mehra <mehra.ashutosh@ibm.com>