Prefer Liberty SCC to OpenJ9 SCC
On some systems trying to use the OpenJ9 SCC leads to
permission denied errors, despite file permissions being
observably correct.

For the time being we'll ignore the OpenJ9 SCC and
continue to build our own SCC layers. In order to do this
we switch to setting/using the OPENJ9_JAVA_OPTIONS
env var in order to override the OpenJ9 image and move
away from IBM_JAVA_OPTIONS, which is deprecated anyway.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>