Use OpenJ9 system SCC where possible (#211)
* Add VERBOSE build arg where needed

Dockerfiles that call configure.sh or populate_scc.sh gate their output
on whether VERBOSE=true or not. Some Dockerfiles have it as a build
arg while others don't. This patch adds it to all such files to be
consistent and make debugging more convenient.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>

* Build on OpenJ9 system SCC when possible

If /opt/java/.scc exists and is readable, assume that an OpenJ9 system
SCC named "openj9_system_scc" exists and build the current image's
SCC layer on top of it. If it doesn't exist or is not readable,
continue to build on the Liberty SCC named "liberty" in
/output/.classCache.

The above will result in images based on IBM JDK continuing to build
a Liberty SCC (as the IBM JDK image does not ship with an SCC) and
OpenJ9 based images building on the OpenJ9 system SCC, except when
directory permissions prevent non-root access.

This last condition is to allow for gracefully handling older OpenJ9
images that did not allow for non-root SCC access.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>