Set OPENJ9_JAVA_OPTIONS env var for containers
Previously OpenJ9 Docker images set OPENJ9_JAVA_OPTIONS to a value
that enabled the use of the system SCC at runtime. Recent OpenJ9 images
have stopped setting it in favour of JAVA_TOOL_OPTIONS, which Liberty is
currently not aware of. This causes the Liberty server script to set
OPENJ9_JAVA_OPTIONS to a default value that does not include the system
SCC.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>