Remove need for sys_ptrace capability at restore time
The sys_ptrace is required at checkpoint time but not at
restore time.  To allow restore time to happen without
the sys_ptrace capability granted to the container
the criu binary is copied to the /opt/criu folder
and not granted the sys_ptrace capability. This
criu copy is used at restore time.