Temporary workaround using -Xgcthreads4
There is an issue restoring gc threads if the host
has more than 4 cores and the JVM tries to increase
the GC thread count on restore.  This option works
around the issue.  The result is we will never have
more than 4 GC threads on a restored server.

This may have some performance issues short-term,
but allows us to make progress on the latest EA
j9 builds.