Add InstantOn to GA images
As of now the Ubuntu semeru images do not have CRIU
support.  Regardless we keep the UBI and Ubuntu images
consistent with our own content.  Once Semeru support is
available for Ubuntu then we will have Liberty pieces in
place to support InstantOn.

For UBI based images the Semeru images only have CRIU
support for openjdk11 and openjdk17 varients.  Support
for openjdk8 is expected later. Similar to Ubuntu images
we are keeping the Liberty content consistent across the
different images with respect to InstantOn.  Once the
Semeru images we are based on get the CRIU support then
we will have the pieces in place across the Liberty image
varients.