Bump PID used for checkpoint to be > 1000
Some environments cause higher PIDs to be used in
the container than 100 which leads to potential
conflicts in the PID needed to restore.

This change increases the checkpoint PID to be > 1000.
More investigation is needed to do this efficiently.
If a more efficient way is found then we may want to
increase this to a higher PID value.  To avoid long
delays during a checkpoint the value of 1000 is being
used for now.