Use instanton.ld.so.cache for checkpointing java
Semeru images may contain a custom InstantOn ld.so.cache that improves
the portability of checkpointed Java programs. Use it by overwriting
ld.so.cache with instanton.ld.so.cache before starting the process being
checkpointed.

Note that in order to overwrite ld.so.cache at checkpoint time we have
to give the checkpoint user write permission to the file; this is done
by making the file group-writable.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>